+++ wies/core/auth_backend.py
@@ -1,6 +1,12 @@
+import logging
+
 from django.contrib.auth.backends import BaseBackend
 
 from .models import Colleague, User
+from .services.events import create_event
+
+
+logger = logging.getLogger(__name__)
 
 
 class AuthBackend(BaseBackend):
@@ -18,6 +24,10 @@ class AuthBackend(BaseBackend):
             user = User.objects.get(email=email)
         except User.DoesNotExist:
             # do not authenticate if User record does not exist
+            logging.info("User not authenticated, user not known.")
+            create_event(email, 'Login.fail', {
+                "reason": "Unknown user",
+            })
             return None
 
         # Link colleague if match on email
@@ -28,6 +38,7 @@ class AuthBackend(BaseBackend):
         except Colleague.DoesNotExist:
             pass
 
+        logging.info("user successfully authenticated")
         return user 
     
     def get_user(self, user_id):